Re fix cemm config (#538)
* 

* revert change in 4d5626d that causes port collision (5005) issue. needs further investigation

* fix permissions issue when using local copy of repository when SELinux is enabled, fix making GPU visible in applicable containers, trim unncessary services for cemm hosting to make easier to follow

* fix PR #516 comment, indentation

* restore the remaining cell whisp model services per #517

* Set more fixed versions to prevent version mismatches during installation

* Fix typo in main.yaml (= instead of == in pip requirements)

* Add missing versioned packages to llava.yaml
* Add missing versioned package to main.yaml

* Fix equal signs in llava.yaml

* Set llava commit to main rather than develop

* Adapt installation for CPU and GPU support on cellwhisp server

* More docs

* Add password to internal dataset

---------

+++ hosting/cemm/cellwhisperer_cycle.sh
@@ -1,3 +1,4 @@
+#!/bin/bash
 # Cycles the CellWhisperer containers on CeMM infrastructure.
 # Due to apparent issues with podman-compose setting up and tearing down networks
 # leading to dns issues we manage the network lifecycle externally.
@@ -6,6 +7,15 @@
 # resolve one another by hostname. Keeping the container environment clean by stopping
 # all the containers and cleaning all the dangling resources with prune seems to 
 # prevent this issue.
+
+# Jump to the directory that contains the script (also contains docker-compose.yaml)
+cd "$(dirname "$0")"
+
+# Make sure that the network filesystem is mounted
+if ! timeout 200 bash -c 'until [ -d /nobackup/lab_bock/projects/cellwhisperer/ ]; do sleep 1; done'; then
+  echo "Error: Directory /nobackup/lab_bock/projects/cellwhisperer/ not found after 200 seconds."
+  exit 1
+fi
 
 # First stop any running cell whisperer containers
 docker compose down
